Rename `metrics::Aggregate` to `metrics::Record`
Move `Record` into its own file.
